updated makesequence to use motifdb
darcs-hash:20070115214023-21d25-83c401c4a6462f35670bd12bfbeba84ff2d4e1cd.gz

+++ dlab/datautils.py
@@ -43,7 +43,7 @@ def offset_add(offsets, data, length=None):
         raise ValueError, "Offset vector must have as many elements as data"
 
     data_len = nx.asarray([len(d) for d in data])
-    offsets = nx.asarray(offsets, dtype='int16')
+    offsets = nx.asarray(offsets, dtype='int32')
 
     stops = data_len + offsets
     length = max(length, stops.max())
